+++ src/Move.java
@@ -1,18 +1,18 @@
 public class Move {
     /** Class describing a move in the game. A TURN will consist of a maximum of four moves, provided they are valid moves. */
-    Move(byte startIndex, byte endIndex) {
+    Move(int startIndex, int endIndex) {
         // FIXME: Use a factory method to save memory.
         this._startIndex = startIndex;
         this._targetIndex = endIndex;
     }
 
     /** Getter for my start index. */
-    public byte start() {
+    public int start() {
         return _startIndex;
     }
 
     /** Getter for my target index. */
-    public byte target() {
+    public int target() {
         return _targetIndex;
     }
 
@@ -22,7 +22,7 @@ public class Move {
     }
 
     /** The starting index of the piece(s) to be moved. */
-    private final byte _startIndex;
+    private final int _startIndex;
     /** The target index of the piece(s) to be moved. */
-    private final byte _targetIndex;
+    private final int _targetIndex;
 }
